test(npm): split test_npm.py into per-module test files
